pyproject.toml
[project]
name = "grpc-server"
version = "0.1.0"
description = "homework qa guru"
authors = [
    {name = "valgoncharov",email = "99090572+valgoncharov@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic (>=2.11.7,<3.0.0)",
    "protobuf (>=6.31.1,<7.0.0)",
    "grpcio-tools (>=1.74.0,<2.0.0)",
    "grpcio-reflection (>=1.74.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
grpcio-tools = "^1.74.0"
